+++ crates/types/src/lib.rs
@@ -18,7 +18,7 @@ mod hashers;
 mod imports;
 mod macros;
 mod non_empty;
-mod nonzero;
+mod non_zero;
 mod query;
 mod response;
 mod result;
@@ -32,7 +32,7 @@ mod utils;
 pub use {
     address::*, app::*, bank::*, binary::*, builder::*, bytearray::*, coin::*, coin_pair::*,
     coins::*, context::*, db::*, denom::*, empty::*, error::*, event::*, hash::*, hashers::*,
-    imports::*, non_empty::*, nonzero::*, query::*, response::*, result::*, serializers::*,
+    imports::*, non_empty::*, non_zero::*, query::*, response::*, result::*, serializers::*,
     signer::*, time::*, tx::*, unique_vec::*, utils::*,
 };
